requirements.txt
requests==2.18.4
pyjwt==1.5.3
websocket-client==0.44.0
pyserial==3.4
